backend/app/requirements.txt
flask
flask-cors
flask-restx
mysql-connector-python